// src/app/Models/IntiativesById.ts
export interface IntiativesDetailsById {
    id?: any,
    title?: any
    description?: any
    imageName?: any
    videoName?: any
    active?: boolean
    publishDate?: any
    puplished?: boolean
    uploadFiles?: boolean
    uploadVideos?: boolean,
    files?: any
    filesVideo?: any
    // isUploadFile?: boolean
   
}